removing transaction_group
