+++ x-pack/plugins/apm/server/lib/service_map/get_service_map_service_node_info.ts
@@ -30,7 +30,7 @@ import {
   percentCgroupMemoryUsedScript,
   percentSystemMemoryUsedScript,
 } from '../metrics/by_agent/shared/memory';
-import { getErrorRate } from '../transaction_groups/get_error_rate';
+import { getErrorRate } from '../transactions/get_error_rate';
 
 interface Options {
   setup: Setup & SetupTimeRange;
