+++ x-pack/plugins/apm/server/lib/transactions/get_transaction_stats/index.ts
@@ -4,11 +4,11 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import { merge } from 'lodash';
-import { TRANSACTION_TYPE } from '../../../common/elasticsearch_fieldnames';
-import { arrayUnionToCallable } from '../../../common/utils/array_union_to_callable';
-import { AggregationInputMap } from '../../../../../typings/elasticsearch';
-import { TransactionGroupRequestBase, TransactionGroupSetup } from './fetcher';
-import { getTransactionDurationFieldForAggregatedTransactions } from '../helpers/aggregated_transactions';
+import { TRANSACTION_TYPE } from '../../../../common/elasticsearch_fieldnames';
+import { arrayUnionToCallable } from '../../../../common/utils/array_union_to_callable';
+import { AggregationInputMap } from '../../../../../../typings/elasticsearch';
+import { TransactionGroupRequestBase, TransactionGroupSetup } from '../fetcher';
+import { getTransactionDurationFieldForAggregatedTransactions } from '../../helpers/aggregated_transactions';
 
 interface MetricParams {
   request: TransactionGroupRequestBase;
