+++ x-pack/plugins/apm/server/lib/transactions/queries.test.ts
@@ -14,6 +14,7 @@ import {
 } from '../../utils/test_helpers';
 // eslint-disable-next-line @kbn/eslint/no-restricted-paths
 import { loggerMock } from '../../../../../../src/core/server/logging/logger.mock';
+import { transactionGroupsFetcher } from './fetcher';
 
 describe('transaction queries', () => {
   let mock: SearchParamsMock;
@@ -112,5 +113,43 @@ describe('transaction queries', () => {
     );
 
     expect(mock.params).toMatchSnapshot();
+  });
+
+  it('fetches top transactions', async () => {
+    const bucketSize = 100;
+    mock = await inspectSearchParams((setup) =>
+      transactionGroupsFetcher(
+        {
+          type: 'top_transactions',
+          serviceName: 'foo',
+          transactionType: 'bar',
+          searchAggregatedTransactions: false,
+        },
+        setup,
+        bucketSize
+      )
+    );
+
+    const allParams = mock.spy.mock.calls.map((call) => call[0]);
+
+    expect(allParams).toMatchSnapshot();
+  });
+
+  it('fetches top traces', async () => {
+    const bucketSize = 100;
+    mock = await inspectSearchParams((setup) =>
+      transactionGroupsFetcher(
+        {
+          type: 'top_traces',
+          searchAggregatedTransactions: false,
+        },
+        setup,
+        bucketSize
+      )
+    );
+
+    const allParams = mock.spy.mock.calls.map((call) => call[0]);
+
+    expect(allParams).toMatchSnapshot();
   });
 });
